Add supporting QCA9377 WiFi chipset in Legato
Modifications:
    1: Add platform adapter code to support QCA9377 WiFi chipset
    2: Change channel data type to support larger channel number used in IEEE802.11a mode
    3: Reorder WiFi hardware start and event monitoring thread creation. QCA9377 WiFi
       driver needs the new order

Resolves: LE-12908
Related: LE-12704

+++ service/daemon/le_wifiAp.c
@@ -126,6 +126,8 @@ void le_wifiAp_RemoveNewEventHandler
  *
  * @return LE_FAULT         The function failed.
  * @return LE_OK            The function succeeded.
+ * @return LE_NOT_FOUND     The WiFi card is absent.
+ * @return LE_NOT_POSSIBLE  The WiFi card may not work.
  *
  */
 //--------------------------------------------------------------------------------------------------
@@ -286,7 +288,7 @@ le_result_t le_wifiAp_SetDiscoverable
 //--------------------------------------------------------------------------------------------------
 le_result_t le_wifiAp_SetChannel
 (
-    int8_t channelNumber
+    uint16_t channelNumber
         ///< [IN]
         ///< the channel number.
 )
